requirements.txt
appdirs==1.4.4
arrow==0.17.0
attrs==20.3.0
bcrypt==3.1.7
blessed==1.17.11
blinker==1.4
blis==0.4.1
breadability==0.1.20
catalogue==1.0.0
certifi==2020.4.5.1
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.8
cymem==2.0.3
dbus-python==1.2.16
decorator==4.4.2
distlib==0.3.1
distro==1.4.0
Django==3.1.2
django-cors-headers==3.2.1
django-filter==2.4.0
django-picklefield==3.0.1
django-q==1.3.4
django-restframework==0.0.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fire==0.3.1
fundamentals==2.2.4
future==0.18.2
gensim==3.8.3
googletrans==2.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.10.20
httpcore==0.9.1
httplib2==0.9.1
httpx==0.13.3
hyperframe==5.2.0
identify==1.5.9
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
iniconfig==1.1.1
jellyfish==0.7.2
joblib==0.15.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
lxml==4.6.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
Morfessor==2.0.6
moviepy==1.0.3
murmurhash==1.0.2
netifaces==0.10.4
networkx==2.4
nltk==3.5
nodeenv==1.5.0
numpy==1.18.4
oauthlib==3.1.0
olefile==0.46
packaging==20.4
pandas==1.0.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.1.2
plac==1.1.3
pluggy==0.13.1
polyglot==16.7.4
polygot==0.1
pre-commit==2.8.2
preshed==3.0.2
profanity-check==1.0.3
proglog==0.1.9
protobuf==3.6.1
psutil==5.7.3
py==1.9.0
pycairo==1.16.2
pycld2==0.41
pycountry==19.8.18
pycups==1.9.73
PyICU==2.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pysbd==0.2.3
pytest==6.1.2
python-dateutil==2.8.1
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
rake-nltk==1.0.4
readability-api==1.0.2
redis==3.5.3
regex==2020.5.14
reportlab==3.5.34
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
rollbar==0.15.1
scikit-learn==0.23.2
scipy==1.4.1
SecretStorage==2.3.1
segtok==1.5.10
simplejson==3.16.0
six==1.15.0
smart-open==3.0.0
sniffio==1.2.0
spacy==2.2.4
sqlparse==0.4.1
srsly==1.0.2
sumy==0.8.1
termcolor==1.1.0
thinc==7.4.0
threadpoolctl==2.1.0
toml==0.10.2
tox==3.20.1
tqdm==4.46.0
translate==3.5.0
Unidecode==1.1.1
urllib3==1.25.11
virtualenv==20.1.0
wadllib==1.3.3
wasabi==0.6.0
wcwidth==0.2.5
wget==3.2
yake==0.3.7
